Do not wait for rewards in localdevnet, avoid ETH sns if not testing
We avoid adding SNs via the ETH network if you just want to spin up
a network to reduce the hardware requirements to run the network. The
ETH sns code path introdces 6 additional SNs to the network. From anecdotal
observations, some of the networking starts to spuriously fail, like,
Pulse handshake bitsets are not getting correctly passed around even
though peers are visible to each other causing block generation to fail.

By spinning up a network without integration tests we can drop the SN's
spawned to 12 nodes again which makes the network more reliable. With the
addition of the storage server also running, this furthermore increases
the hardware requirements for this private network to run.